requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.1
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
clear-cache==1.0.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==3.2
django-3-jet==1.0.8
django-admin-autocomplete-filter==0.7.1
django-adminlte-3==0.1.6
django-adminlte3==0.1.6
django-bootstrap3==22.1
django-easy-select2==1.5.8
django-filter==22.1
django-import-export==2.8.0
django-widget-tweaks==1.4.8
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.7.1
fpdf==1.7.2
future==0.18.2
html5lib==1.1
idna==3.3
lxml==4.8.0
MarkupPy==1.14
numpy==1.23.4
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
pandas==1.5.1
Pillow==9.1.0
platformdirs==2.5.2
pycparser==2.21
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyPDF2==2.10.5
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
pytube==12.1.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
razorpay==1.3.0
reportlab==3.6.9
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tablib==3.2.1
tinycss2==1.1.1
tqdm==4.64.0
typing_extensions==4.4.0

tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
virtualenv==20.14.1
webencodings==0.5.1
xhtml2pdf==0.2.7
xlrd==2.0.1
xlwt==1.3.0
num2words==0.5.12
mysqlclient==2.0.1
wheel
django-admin-rangefilter
django-admin-list-filter-dropdown